Issue #2 - Moved weapon_type to new header file.
